Continued working on machine's window.
Changed how the rendering works, moved the handling from inside WNDPROC,
to outside the message loop. the loop now uses PeekMessage instead of
GetMessage, since the later blocks execution.
Added a callback to the window loop, but this will be changed probably.

Fixed a mistake with setting up the bitmap headers -
for top down buffer, the height needed to be negative number.

Machine command has some POC code to checks if the window and
rendering works properly.