Machine command changes
Cleaned up the code a little bit, and preparing to integrate with the
chip8 virtual machine code.

Something to think about, maybe during this integration
I should also start with adding support for the debugger api.
Because it will make life easier to debug and fix bugs in the
current implementation of the virtual machine.