Draw executor - WIP
Fixing the calculation of the draw executor - current state is still in
the middle of changing and fixing the calculations, so those are still
wrong.

Another missing feature from the draw executor is the ability to wrap
sprite in case of out of bound drawing (currently implementation will
cause the sprite to go down a row) - this will be added later.