Implement freeze mechanism for wait for key press - WIP.
While trying different ROMS, there seems to still be issues with
drawing. With the current state of the machine, it is
extremely hard to debug, therefor I decided the focus now should be on
creating a debug interface and assembler and also to improve the
implementation of the dissasembler to support more verbose output.