Refactored disasm - API changes
Changed the disassembler from just returning a string, to returning a
struct describing an instruction. The code still doesn't support the
functionality of changing output format and generating labels.

In the middle of refactoring the disassembler, I found
misimplementation of the machine instructions 0xFx55 and 0xFx65,
the previous implementation didn't take into account partial storing of
V registers, and just always saved all of them into memory. - This was
the origin of the issue seen with drawing pixels.